// x-pack/solutions/security/plugins/security_solution/common/api/entity_analytics/risk_engine/preview_matched_entities_route.gen.ts
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License
 * 2.0; you may not use this file except in compliance with the Elastic License
 * 2.0.
 */

/*
 * NOTICE: Do not edit this file manually.
 * This file is automatically generated by the OpenAPI Generator, @kbn/openapi-generator.
 *
 * info:
 *   title: Risk Scoring API
 *   version: 1
 */

import { z } from '@kbn/zod';

import {
  DataViewId,
  Filter,
  PageSize,
  IdentifierType,
  DateRange,
  RiskScoreWeights,
  EntityRiskScoreRecord,
} from '../common/common.gen';

export type MatchedPreviewRiskScoreRequest = z.infer<typeof MatchedPreviewRiskScoreRequest>;
export const MatchedPreviewRiskScoreRequest = z.object({
  /**
   * The list of entity names which are the same
   */
  matched_entities: z.array(z.string()),
  /**
   * The identifier of the Kibana data view to be used when generating risk scores. If a data view is not found, the provided ID will be used as the query's index pattern instead.
   */
  data_view_id: DataViewId,
  /**
   * If set to `true`, a `debug` key is added to the response, containing both the internal request and response with elasticsearch.
   */
  debug: z.boolean().optional(),
  /**
   * An elasticsearch DSL filter object. Used to filter the data being scored, which implicitly filters the risk scores returned.
   */
  filter: Filter.optional(),
  page_size: PageSize.optional(),
  /**
   * Used to restrict the type of risk scores involved. If unspecified, both `host` and `user` scores will be returned.
   */
  identifier_type: IdentifierType.optional(),
  /**
   * Defines the time period over which scores will be evaluated. If unspecified, a range of `[now, now-30d]` will be used.
   */
  range: DateRange.optional(),
  weights: RiskScoreWeights.optional(),
  /**
   * A list of alert statuses to exclude from the risk score calculation. If unspecified, all alert statuses are included.
   */
  exclude_alert_statuses: z.array(z.string()).optional(),
  /**
   * A list of alert tags to exclude from the risk score calculation. If unspecified, all alert tags are included.
   */
  exclude_alert_tags: z.array(z.string()).optional(),
});

export type MatchedPreviewRiskScoreResponse = z.infer<typeof MatchedPreviewRiskScoreResponse>;
export const MatchedPreviewRiskScoreResponse = z.object({
  /**
   * Object containing debug information, particularly the internal request and response from elasticsearch
   */
  debug: z
    .object({
      request: z.string().optional(),
      response: z.string().optional(),
    })
    .optional(),
  scores: z.object({
    /**
     * A list of host risk scores
     */
    host: z.array(EntityRiskScoreRecord).optional(),
    /**
     * A list of user risk scores
     */
    user: z.array(EntityRiskScoreRecord).optional(),
    /**
     * A list of service risk scores
     */
    service: z.array(EntityRiskScoreRecord).optional(),
    /**
     * A list of generic entities risk scores
     */
    generic: z.array(EntityRiskScoreRecord).optional(),
  }),
});

export type MatchedEntitiesPreviewRiskScoreRequestBody = z.infer<
  typeof MatchedEntitiesPreviewRiskScoreRequestBody
>;
export const MatchedEntitiesPreviewRiskScoreRequestBody = MatchedPreviewRiskScoreRequest;
export type MatchedEntitiesPreviewRiskScoreRequestBodyInput = z.input<
  typeof MatchedEntitiesPreviewRiskScoreRequestBody
>;

export type MatchedEntitiesPreviewRiskScoreResponse = z.infer<
  typeof MatchedEntitiesPreviewRiskScoreResponse
>;
export const MatchedEntitiesPreviewRiskScoreResponse = MatchedPreviewRiskScoreResponse;